OAK-10280: Occasional failure to start docker container
Consolidate multiple copies of AzuriteDockerRule
Skip test when waiting for port or startup message times out
