+++ oak-segment-azure/src/test/java/org/apache/jackrabbit/oak/segment/azure/journal/AzureTarRevisionsTest.java
@@ -17,8 +17,9 @@
 package org.apache.jackrabbit.oak.segment.azure.journal;
 
 import com.microsoft.azure.storage.blob.CloudBlobContainer;
+
+import org.apache.jackrabbit.oak.blob.cloud.azure.blobstorage.AzuriteDockerRule;
 import org.apache.jackrabbit.oak.segment.spi.persistence.SegmentNodeStorePersistence;
-import org.apache.jackrabbit.oak.segment.azure.AzuriteDockerRule;
 import org.apache.jackrabbit.oak.segment.file.TarRevisionsTest;
 import org.apache.jackrabbit.oak.segment.azure.AzurePersistence;
 import org.junit.Before;
