+++ oak-upgrade/pom.xml
@@ -215,7 +215,7 @@
     </dependency>
     <dependency>
       <groupId>org.apache.jackrabbit</groupId>
-      <artifactId>oak-segment-azure</artifactId>
+      <artifactId>oak-blob-cloud-azure</artifactId>
       <version>${project.version}</version>
       <type>test-jar</type>
       <scope>test</scope>
